Added mock tests for service layer

+++ storage-service/src/main/java/by/itacademy/javaenterprise/knyazev/services/CategoriesService.java
@@ -1,6 +1,7 @@
 package by.itacademy.javaenterprise.knyazev.services;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -19,16 +20,25 @@ import by.itacademy.javaenterprise.knyazev.services.exceptions.ServiceException;
 @Service
 public class CategoriesService {
 	private static final Logger logger = LoggerFactory.getLogger(CategoriesService.class);
-	@Autowired
 	private CategoriesDAO categoriesDAO;
+	private GoodsDAO goodsDAO;
+
 	@Autowired
-	private GoodsDAO goodsDAO;	
+	public CategoriesService(CategoriesDAO categoriesDAO, GoodsDAO goodsDAO) {
+		this.categoriesDAO = categoriesDAO;
+		this.goodsDAO = goodsDAO;
+	}
 
 	@Transactional
-	public Map<Category, Long> showAll() {
+	public Map<Category, Long> showAll() throws ServiceException {
 		Map<Category, Long> categoriesGoodsQuantity = new HashMap<>();
 
-		categoriesDAO.findAll().stream().forEach(c -> {
+		List<Category> categories = categoriesDAO.findAll();
+
+		if (categories == null || categories.isEmpty())
+			throw new ServiceException("Categories not found!");
+
+		categories.stream().forEach(c -> {
 			categoriesGoodsQuantity.put(c, goodsDAO.findCountByCategoryId(c.getId()));
 		});
 
@@ -37,13 +47,17 @@ public class CategoriesService {
 
 	@Transactional
 	public Category showCategoryById(Long id) throws ServiceException {
+		if (id == null || id <= 0l) {
+			logger.error("Error loading category, id must be not {}", id);
+			throw new ServiceException("Error loading category, id must be not " + id + "!");
+		}
 
 		Optional<Category> category = categoriesDAO.findById(id);
 		if (category.isPresent()) {
 			category.get().getGoods().size();
 			return category.get();
 		}
-		
+
 		logger.error("Error loading category on id={}!", id);
 		throw new ServiceException("Error loading category on id=" + id + "!");
 	}
@@ -54,8 +68,8 @@ public class CategoriesService {
 		Category existing = categoriesDAO.findByName(category.getName());
 
 		if (existing != null) {
-			logger.error("Category with name={} already exists!", category.getName());
-			throw new ServiceException("Category with name=" + category.getName() + " already exists!");
+			logger.error("Can't save category. Category with name={} already exists!", category.getName());
+			throw new ServiceException("Can't save category. Category with name=" + category.getName() + " already exists!");
 		}
 
 		Category savedCategory = null;
@@ -64,35 +78,47 @@ public class CategoriesService {
 		}
 
 		if (savedCategory != null) {
-			if (savedCategory.getId() != null && (savedCategory.getId() > 0L))
+			if ((savedCategory.getId() != null) && (savedCategory.getId() > 0L)) {
 				return savedCategory;
+			}
 		}
-		
+
 		logger.error("Error on saving category!");
 		throw new ServiceException("Error on saving category!");
 	}
 
 	@Transactional
 	public Category updateCategory(Category category) throws ServiceException {
-		if (category.getId() == null || category.getId() <= 0L) {
+		Long id = (category != null) ? category.getId() : null;
+		
+		if (id == null || id <= 0L) {
 			logger.error("Category must have a valid id for updating!");
 			throw new ServiceException("Category must have a valid id for updating!");
 		}
-		
-		if (categoriesDAO.existsById(category.getId())) {
-			return categoriesDAO.save(category);
+
+		if (!categoriesDAO.existsById(category.getId())) {
+			logger.error("Error on updating category by non existing id = {}!", id);
+			throw new ServiceException("Error on updating category by non existing id = " + id);
 		}
 		
+		Category updatedCategory = categoriesDAO.save(category);
+		
+		if (updatedCategory != null) {
+			if ((updatedCategory.getId() != null) && (updatedCategory.getId() > 0L)) {
+				return updatedCategory;
+			}
+		}
+
 		logger.error("Error on updating category!");
 		throw new ServiceException("Error on updating category!");
 	}
-	
+
 	@Transactional
 	public void deleteCategory(Long id) throws ServiceException {
-				
-		if (categoriesDAO.existsById(id)) {
+
+		if ((id != null) && (categoriesDAO.existsById(id))) {
 			categoriesDAO.deleteById(id);
-		} else {	
+		} else {
 			logger.error("Error on deleting category on id={}!", id);
 			throw new ServiceException("Error on deleting category on id=" + id + "!");
 		}
