+++ storage-service/src/main/java/by/itacademy/javaenterprise/knyazev/services/GoodsService.java
@@ -18,9 +18,12 @@ import by.itacademy.javaenterprise.knyazev.services.exceptions.ServiceException;
 @Service
 public class GoodsService {
 	private static final Logger logger = LoggerFactory.getLogger(GoodsService.class);
-	@Autowired
 	private GoodsDAO goodsDAO;
-	
+
+	@Autowired
+	public GoodsService(GoodsDAO goodsDAO) {
+		this.goodsDAO = goodsDAO;
+	}
 
 	@Transactional
 	public List<Good> showAll() {
@@ -29,22 +32,29 @@ public class GoodsService {
 
 	@Transactional
 	public List<Good> showAll(Integer page, Integer size) throws ServiceException {
-		int realPage = page.intValue() - 1;
+		if (page == null || size == null || page < 1L || size < 1L) {
+			throw new ServiceException(
+					"Page must be equals or greater then 1 and size must be equals or greater than 1 for showing all goods");
+		}
+
+		int realPage = page - 1;
 		List<Good> goods = goodsDAO.findAll(PageRequest.of(realPage, size)).getContent();
 
 		if (!goods.isEmpty()) {
 			return goods;
 		}
 
-		logger.error("Error in page mumber");
-		throw new ServiceException("Error in page mumber");
+		logger.error("Error on loading goods");
+		throw new ServiceException("Error on loading goods");
 	}
 
 	@Transactional
 	public Good showGoodById(Long id) throws ServiceException {
-		Optional<Good> good = goodsDAO.findById(id);
-		if (good.isPresent()) {
-			return good.get();
+		if (id != null) {
+			Optional<Good> good = goodsDAO.findById(id);
+			if (good.isPresent()) {
+				return good.get();
+			}
 		}
 		logger.error("Error loading good on id={}", id);
 		throw new ServiceException("Error loading good on id=" + id + "!");
@@ -53,32 +63,29 @@ public class GoodsService {
 	@Transactional
 	public Good saveGood(Good good) throws ServiceException {
 
-		if (good.getId() != null) {
-			logger.error("Good id must be equals to null for saving!");
-			throw new ServiceException("Good id must be equals to null for saving!");
+		if (good == null || good.getId() != null) {
+			logger.error("Good have to exists and good id must be equals to null for saving!");
+			throw new ServiceException("Good have to exists and good id must be equals to null for saving!");
 		}
 
-		List<Good> existingGoods = goodsDAO.findByNameAndSort(good.getName(), good.getSort());
-		Long categoryId = good.getCategory().getId();
-		Long producerId = good.getProducer().getId();
+		String name = good.getName();
+		String sort = good.getSort();
+		Long categoryId = (good.getCategory() != null) ? good.getCategory().getId() : null;
+		Long producerId = (good.getProducer() != null) ? good.getProducer().getId() : null;
 
-		if (!existingGoods.isEmpty()) {
-			for (Good existingGood : existingGoods) {
+		Good existingGood = goodsDAO.findByNameAndSortAndCategoryIdAndProducerId(name, sort, categoryId, producerId);
 
-				if ((categoryId == existingGood.getCategory().getId())
-						&& (producerId == existingGood.getProducer().getId())) {
-					logger.error("Saving good already exists with unique name, sort, category and producer!");
-					throw new ServiceException(
-							"Saving good already exists with unique name, sort, category and producer!");
-				}
-			}
+		if (existingGood != null) {
+			logger.error("Saving good already exists with unique name, sort, category and producer!");
+			throw new ServiceException("Saving good already exists with unique name, sort, category and producer!");
 		}
 
 		Good savedGood = goodsDAO.save(good);
 
 		if (savedGood != null) {
-			if (savedGood.getId() != null && (savedGood.getId() > 0L))
+			if ((savedGood.getId() != null) && (savedGood.getId() > 0L)) {
 				return savedGood;
+			}
 		}
 		logger.error("Error on saving good!");
 		throw new ServiceException("Error on saving good!");
@@ -86,25 +93,38 @@ public class GoodsService {
 
 	@Transactional
 	public Good updateGood(Good good) throws ServiceException {
-		if (good.getId() == null || good.getId() <= 0L) {
+		Long id = (good != null) ? good.getId() : null;
+		if (id == null || id <= 0L) {
 			logger.error("Good must have a valid id for updating!");
 			throw new ServiceException("Good must have a valid id for updating!");
 		}
-
-		if (goodsDAO.existsById(good.getId())) {
-			return goodsDAO.save(good);
+		
+		
+		if (!goodsDAO.existsById(id)) {
+			logger.error("Error on updating good by non existing id = {}!", id);
+			throw new ServiceException("Error on updating good by non existing id = " + id);
 		}
-		logger.error("Error on updating category!");
-		throw new ServiceException("Error on updating category!");
+		
+		Good updatedGood = goodsDAO.save(good);
+		
+		if (updatedGood != null) {
+			if ((updatedGood.getId() != null) && (updatedGood.getId() > 0L)) {
+				return updatedGood;
+			}
+		}
+		
+		logger.error("Error on updating good!");
+		throw new ServiceException("Error on updating good!");
 	}
 
 	@Transactional
 	public void deleteGood(Long id) throws ServiceException {
-		if (goodsDAO.existsById(id)) {
+		
+		if ((id != null) && (goodsDAO.existsById(id))) {
 			goodsDAO.deleteById(id);
 		} else {
-			logger.error("Error on deleting category on id={}!", id);
-			throw new ServiceException("Error on deleting category on id=" + id + "!");
+			logger.error("Error on deleting good on id={}!", id);
+			throw new ServiceException("Error on deleting good on id=" + id + "!");
 		}
 	}
